Refactors `pypi_api.py`
- Breaks out code that can be common to other API implementations
- Rearranges the unit tests accordingly.
- General updates to our latest style choices
