+++ anaconda_packaging_utils/api/_types.py
@@ -1,12 +1,22 @@
 """
 File:           _types.py
-Description:    Contains types and classes used internally for the API modules.
+Description:    Contains private constants, types, and classes used in the APIs provided by this package.
 """
+
+
+from typing import Final
+
+#### Constants ####
+
+# Timeout of HTTP requests, in seconds
+DEFAULT_HTTP_REQ_TIMEOUT: Final[int] = 60
+
+#### Classes ####
 
 
 class BaseApiException(Exception):
     """
-    Generic exception indicating an unrecoverable failure of this API.
+    Base generic API exception indicating an unrecoverable failure of this API.
 
     This exception is meant to condense many possible failures into one generic error. The thinking is, if the calling
     code runs into any API failure, there isn't much that can be done. So it is easier for the caller to handle one
